drivers: lora: sx126x: Support software-managed RF switches
The SX126x supports using DIO2 to control an TX switch, but only if
the switch can be operated using a single control signal. Add support
for RF switches that are wired to a GPIO instead of the radio chip
itself. This makes it possible to use RF switches that require two
control signals (one for the RX port and one for the TX port) by
wiring them to two GPIOs on the MCU.

RF switch control is implementing by tracking the commands sent to the
radio and the interrupts received from the radio.

Signed-off-by: Andreas Sandberg <andreas@sandberg.pp.se>